Move component & hooks to drafts folder
