+++ src/drafts/InlineAutocomplete/_AutocompleteSuggestions.tsx
@@ -1,9 +1,9 @@
 import React, {useCallback, useState} from 'react'
-import {Spinner} from '..'
-import {ActionList, ActionListItemProps} from '../ActionList'
-import Box from '../Box'
-import {ComboboxCommitEvent, useCombobox} from '../hooks'
-import Overlay from '../Overlay'
+import {Spinner} from '../..'
+import {ActionList, ActionListItemProps} from '../../ActionList'
+import Box from '../../Box'
+import {ComboboxCommitEvent, useCombobox} from '../hooks/useCombobox'
+import Overlay from '../../Overlay'
 
 import {Suggestion, Suggestions, TextInputElement} from './types'
 import {getSuggestionKey, getSuggestionValue} from './utils'
