+++ src/drafts/index.ts
@@ -8,11 +8,11 @@ export * from '../NavList'
 export * from '../Dialog/Dialog'
 export * from '../SegmentedControl'
 
-export {default as InlineAutocomplete} from '../InlineAutocomplete'
+export {default as InlineAutocomplete} from './InlineAutocomplete'
 export type {
   InlineAutocompleteProps,
   ShowSuggestionsEvent,
   Suggestion,
   Suggestions,
   Trigger
-} from '../InlineAutocomplete'
+} from './InlineAutocomplete'
